Exception handling

+++ dscommerce/src/main/java/com/weildev/dscommerce/services/ProductService.java
@@ -3,12 +3,16 @@ package com.weildev.dscommerce.services;
 import com.weildev.dscommerce.dto.ProductDTO;
 import com.weildev.dscommerce.entities.Product;
 import com.weildev.dscommerce.repositories.ProductRepository;
+import com.weildev.dscommerce.services.exceptions.DatabaseException;
 import com.weildev.dscommerce.services.exceptions.ResourceNotFoundException;
+import jakarta.persistence.EntityNotFoundException;
 import org.springframework.beans.BeanUtils;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Propagation;
 import org.springframework.transaction.annotation.Transactional;
 
 @Service
@@ -41,15 +45,26 @@ public class ProductService {
 
     @Transactional
     public ProductDTO update(Long id, ProductDTO productDTO) {
-        Product entity = productRepository.getReferenceById(id);
-        BeanUtils.copyProperties(productDTO, entity, "id");
+        try{
+            Product entity = productRepository.getReferenceById(id);
+            BeanUtils.copyProperties(productDTO, entity, "id");
 
-        entity = productRepository.save(entity);
-        return new ProductDTO(entity);
+            entity = productRepository.save(entity);
+            return new ProductDTO(entity);
+        }catch (EntityNotFoundException e) {
+            throw new ResourceNotFoundException("Resource not found");
+        }
     }
 
-    @Transactional
+    @Transactional(propagation = Propagation.SUPPORTS)
     public void delete(Long id) {
-        productRepository.deleteById(id);
+        if (!productRepository.existsById(id)) {
+            throw new ResourceNotFoundException("Resource not found");
+        }
+        try {
+            productRepository.deleteById(id);
+        }catch (DataIntegrityViolationException e) {
+            throw new DatabaseException("Referential integrity failure");
+        }
     }
 }
